refactor(renderers): split render.ts, add withDedup + shared inputs
render.ts (795 lines) split into:
  context-builder.ts — createRenderContext, cache wrapping, pricing,
    preview mode, PreparedRender type
  flatten.ts — flattenClips, HoistedCaption, DeferredAudio types
  render.ts — thin orchestrator (~400 lines, was 795)

withDedup helper (dedup.ts) eliminates 30 lines of copy-paste from
image/video/speech/music renderers:
  - pre-resolved short-circuit
  - pendingFiles dedup by cache key
  - progress task lifecycle (addTask/startTask/completeTask)

inputs.ts consolidates resolveImageInput, resolveVideoInput,
resolveAudioInput — previously duplicated between image.ts and video.ts.

resolveCacheStorage moved from render.ts (parent) to context-builder.ts,
eliminating the last duplicate.
