+++ src/react/renderers/music.ts
@@ -3,14 +3,13 @@ import type { generateMusic } from "../../ai-sdk/generate-music";
 import { ResolvedElement } from "../resolved-element";
 import type { MusicProps, VargElement } from "../types";
 import type { RenderContext } from "./context";
-import { addTask, completeTask, startTask } from "./progress";
+import { withDedup } from "./dedup";
 import { computeCacheKey } from "./utils";
 
 export async function renderMusic(
   element: VargElement<"music">,
   ctx: RenderContext,
 ): Promise<File> {
-  // If already resolved via `await Music(...)`, reuse the pre-generated file
   if (element instanceof ResolvedElement) {
     ctx.generatedFiles.push(element.meta.file);
     return element.meta.file;
@@ -24,30 +23,16 @@ export async function renderMusic(
     throw new Error("Music requires prompt and model (or set defaults.music)");
   }
 
+  const modelId = model.modelId ?? "music";
   const cacheKey = computeCacheKey(element);
-  const cacheKeyStr = JSON.stringify(cacheKey);
 
-  // Deduplicate concurrent renders of the same music element
-  const pendingRender = ctx.pendingFiles.get(cacheKeyStr);
-  if (pendingRender) {
-    return pendingRender;
-  }
-
-  const renderPromise = (async () => {
-    const modelId = model.modelId ?? "music";
-    const taskId = ctx.progress
-      ? addTask(ctx.progress, "music", modelId)
-      : null;
-    if (taskId && ctx.progress) startTask(ctx.progress, taskId);
-
+  return withDedup(element, ctx, "music", modelId, async () => {
     const { audio } = await ctx.generateMusic({
       model,
       prompt,
       duration: props.duration,
       cacheKey,
     } as Parameters<typeof generateMusic>[0]);
-
-    if (taskId && ctx.progress) completeTask(ctx.progress, taskId);
 
     if (!audio?.uint8Array) {
       throw new Error(
@@ -70,12 +55,6 @@ export async function renderMusic(
       prompt,
     });
 
-    ctx.generatedFiles.push(file);
-
     return file;
-  })();
-
-  ctx.pendingFiles.set(cacheKeyStr, renderPromise);
-
-  return renderPromise;
+  });
 }
