+++ src/react/renderers/render.ts
@@ -1,21 +1,16 @@
-import type { ImageModelV3 } from "@ai-sdk/provider";
-import {
-  generateImage,
-  experimental_generateSpeech as generateSpeech,
-  wrapImageModel,
-} from "ai";
+/**
+ * Compose phase: walk the (already materialized) tree, assemble the
+ * timeline, and run editly + captions burn-in.
+ *
+ * When called via `render()`, every generable element has been executed by
+ * the plan executor and carries `meta` — the per-element renderers inside
+ * only short-circuit on pre-resolved files. Calling renderRoot directly
+ * (without a prepared context) still works: renderers generate on demand,
+ * preserving the legacy recursive behavior.
+ */
+
 import pMap from "p-map";
-import { type CacheStorage, withCache } from "../../ai-sdk/cache";
-import type { File, File as VargFile } from "../../ai-sdk/file";
-import { fileCache } from "../../ai-sdk/file-cache";
-import { generateMusic } from "../../ai-sdk/generate-music";
-import { generateVideo } from "../../ai-sdk/generate-video";
-import {
-  imagePlaceholderFallbackMiddleware,
-  placeholderFallbackMiddleware,
-  wrapVideoModel,
-} from "../../ai-sdk/middleware";
-import { editly, localBackend } from "../../ai-sdk/providers/editly";
+import { editly } from "../../ai-sdk/providers/editly";
 import type {
   AudioTrack,
   Clip,
@@ -27,7 +22,6 @@ import type {
   ClipProps,
   MusicProps,
   OverlayProps,
-  RenderMode,
   RenderOptions,
   RenderProps,
   RenderResult,
@@ -38,17 +32,13 @@ import { audioMixVolume, renderAudio } from "./audio";
 import { burnCaptions } from "./burn-captions";
 import { type CaptionsResult, renderCaptions } from "./captions";
 import { renderClip } from "./clip";
-import type { RenderContext } from "./context";
+import { createRenderContext, type PreparedRender } from "./context-builder";
 import type { EmojiOverlay } from "./emoji";
+import { type FlattenResult, flattenClips } from "./flatten";
 import { renderImage } from "./image";
 import { mergeAssFiles, shiftAssTimestamps } from "./merge-ass";
 import { renderMusic } from "./music";
-import {
-  addTask,
-  completeTask,
-  createProgressTracker,
-  startTask,
-} from "./progress";
+import { addTask, completeTask, startTask } from "./progress";
 import { renderSpeech } from "./speech";
 import { resolvePath } from "./utils";
 import { renderVideo } from "./video";
@@ -59,361 +49,37 @@ interface RenderedOverlay {
   isVideo: boolean;
 }
 
-function resolveCacheStorage(
-  cache: string | CacheStorage | undefined,
-): CacheStorage | undefined {
-  if (!cache) return undefined;
-  if (typeof cache === "string") {
-    return fileCache({ dir: cache });
-  }
-  return cache;
-}
-
-function toImageModelV3(
-  model: Parameters<typeof generateImage>[0]["model"],
-): ImageModelV3 {
-  if (typeof model === "object" && model.specificationVersion === "v3") {
-    return model;
-  }
-  // for string IDs and v2 models, create a shell that satisfies the type.
-  // in preview mode the middleware intercepts before doGenerate is called.
-  const modelId = typeof model === "string" ? model : model.modelId;
-  return {
-    specificationVersion: "v3",
-    provider: "placeholder",
-    modelId,
-    maxImagesPerCall: 1,
-    doGenerate: async () => {
-      throw new Error(
-        `toImageModelV3 shell: doGenerate should not be called in preview mode (model: ${modelId})`,
-      );
-    },
-  };
-}
-
-/** RenderContext plus the render-scoped state renderRoot needs alongside it. */
-export interface PreparedRender {
-  ctx: RenderContext;
-  progress: ReturnType<typeof createProgressTracker>;
-  mode: RenderMode;
-  placeholderCount: { images: number; videos: number; total: number };
-}
-
 /**
- * Build the RenderContext (cached+wrapped generators, backend, progress)
- * for a composition. Shared by the plan executor and the compose phase so
- * both see the same pendingFiles dedup map, progress tracker, and
- * generatedFiles list.
+ * Walk <Render> children, flatten clips, collect overlays/music/audio.
  */
-export function createRenderContext(
+async function collectTopLevel(
   element: VargElement<"render">,
-  options: RenderOptions,
-): PreparedRender {
-  const props = element.props as RenderProps;
-  const progress = createProgressTracker(options.quiet ?? false);
-
-  const mode: RenderMode = options.mode ?? "strict";
-  const placeholderCount = { images: 0, videos: 0, total: 0 };
-
-  const trackPlaceholder = (type: "image" | "video") => {
-    placeholderCount[type === "image" ? "images" : "videos"]++;
-    placeholderCount.total++;
-  };
-
-  const cacheStorage = resolveCacheStorage(options.cache);
-
-  const cachedGenerateImage = cacheStorage
-    ? withCache(generateImage, { storage: cacheStorage })
-    : generateImage;
-
-  const cachedGenerateVideo = cacheStorage
-    ? withCache(generateVideo, { storage: cacheStorage })
-    : generateVideo;
-
-  const cachedGenerateSpeech = cacheStorage
-    ? withCache(generateSpeech, { storage: cacheStorage })
-    : generateSpeech;
-
-  const cachedGenerateMusic = cacheStorage
-    ? withCache(generateMusic, { storage: cacheStorage })
-    : generateMusic;
-
-  const onGeneration = options.onGeneration;
-
-  /** Extract pricing metadata from provider response and emit via callback. */
-  // biome-ignore lint/suspicious/noExplicitAny: result shapes vary across AI SDK model types
-  const emitPricing = (
-    type: "image" | "video" | "speech" | "music",
-    modelId: string,
-    result: any,
-  ) => {
-    if (!onGeneration) return;
-    const vargMeta = result?.providerMetadata?.varg as
-      | { pricing?: Record<string, unknown>; jobId?: string }
-      | undefined;
-    if (vargMeta?.pricing) {
-      const p = vargMeta.pricing;
-      onGeneration({
-        type,
-        model: modelId,
-        estimated: p.estimated as number | undefined,
-        actual: p.actual as number | undefined,
-        billing: p.billing as "metered" | "byok" | "x402" | undefined,
-        cached: p.cached as boolean | undefined,
-        jobId: vargMeta.jobId,
-      });
-    }
-  };
-
-  const wrapGenerateImage: typeof generateImage = async (opts) => {
-    if (mode === "preview") {
-      trackPlaceholder("image");
-      return cachedGenerateImage({
-        ...opts,
-        model: wrapImageModel({
-          model: toImageModelV3(opts.model),
-          middleware: imagePlaceholderFallbackMiddleware({
-            mode: "preview",
-            onFallback: () => {},
-          }),
-        }),
-        skipCacheWrite: true,
-      } as Parameters<typeof cachedGenerateImage>[0]);
-    }
-
-    const result = await cachedGenerateImage(opts);
-    const imgModelId =
-      typeof opts.model === "string" ? opts.model : opts.model.modelId;
-    emitPricing("image", imgModelId, result);
-    return result;
-  };
-
-  const wrapGenerateVideo: typeof generateVideo = async (opts) => {
-    if (mode === "preview") {
-      trackPlaceholder("video");
-      return cachedGenerateVideo({
-        ...opts,
-        model: wrapVideoModel({
-          model: opts.model,
-          middleware: placeholderFallbackMiddleware({
-            mode: "preview",
-            onFallback: () => {},
-          }),
-        }),
-        skipCacheWrite: true,
-      } as Parameters<typeof cachedGenerateVideo>[0]);
-    }
-
-    const result = await cachedGenerateVideo(opts);
-    emitPricing("video", opts.model.modelId, result);
-    return result;
-  };
-
-  const wrapGenerateSpeech: typeof generateSpeech = async (opts) => {
-    const result = await cachedGenerateSpeech(opts);
-    const speechModelId =
-      typeof opts.model === "string" ? opts.model : opts.model.modelId;
-    emitPricing("speech", speechModelId, result);
-    return result;
-  };
-
-  const wrapGenerateMusic: typeof generateMusic = async (opts) => {
-    const result = await cachedGenerateMusic(opts);
-    const musicModelId =
-      typeof opts.model === "string" ? opts.model : opts.model.modelId;
-    emitPricing("music", musicModelId, result);
-    return result;
-  };
-
-  const backend = options.backend ?? localBackend;
-  const tempFiles: string[] = [];
-  const generatedFiles: VargFile[] = [];
-  const ctx: RenderContext = {
-    width: props.width ?? 1920,
-    height: props.height ?? 1080,
-    fps: props.fps ?? 30,
-    cache: cacheStorage,
-    storage: options.storage,
-    generateImage: wrapGenerateImage,
-    generateVideo: wrapGenerateVideo,
-    generateSpeech: onGeneration ? wrapGenerateSpeech : cachedGenerateSpeech,
-    generateMusic: onGeneration ? wrapGenerateMusic : cachedGenerateMusic,
-    tempFiles,
-    progress,
-    pendingFiles: new Map<string, Promise<File>>(),
-    defaults: options.defaults,
-    backend,
-    generatedFiles,
-  };
-
-  return { ctx, progress, mode, placeholderCount };
-}
-
-/**
- * Compose phase: walk the (already materialized) tree, assemble the
- * timeline, and run editly + captions burn-in.
- *
- * When called via `render()`, every generable element has been executed by
- * the plan executor and carries `meta` — the per-element renderers inside
- * only short-circuit on pre-resolved files. Calling renderRoot directly
- * (without a prepared context) still works: renderers generate on demand,
- * preserving the legacy recursive behavior.
- */
-export async function renderRoot(
-  element: VargElement<"render">,
-  options: RenderOptions,
-  prepared?: PreparedRender,
-): Promise<RenderResult> {
-  const props = element.props as RenderProps;
-  const { ctx, progress, mode, placeholderCount } =
-    prepared ?? createRenderContext(element, options);
-  const generatedFiles = ctx.generatedFiles;
-
-  const clipElements: VargElement<"clip">[] = [];
+  ctx: ReturnType<typeof createRenderContext>["ctx"],
+): Promise<{
+  flatten: FlattenResult;
+  overlayElements: VargElement<"overlay">[];
+  musicElements: VargElement<"music">[];
+  audioTracks: AudioTrack[];
+  hoistedCaptions: FlattenResult["hoistedCaptions"];
+}> {
   const overlayElements: VargElement<"overlay">[] = [];
   const musicElements: VargElement<"music">[] = [];
   const audioTracks: AudioTrack[] = [];
 
-  // ---------------------------------------------------------------------------
-  // Hoisted captions: track which clip each caption came from so we can apply
-  // the correct timeline offset when stitching audio and ASS files.
-  // ---------------------------------------------------------------------------
-  interface HoistedCaption {
-    element: VargElement<"captions">;
-    clipIndex: number;
-  }
-  const hoistedCaptions: HoistedCaption[] = [];
-
-  // ---------------------------------------------------------------------------
-  // Deferred audio: speech/music at a container-clip level that needs to be
-  // offset to the container's start position in the timeline.
-  // We don't know the offset yet (clips haven't been laid out), so we record
-  // the clipIndex of the first leaf clip inside the container and resolve the
-  // offset later, after clipStartOffsets are computed.
-  // ---------------------------------------------------------------------------
-  interface DeferredAudio {
-    element:
-      | VargElement<"speech">
-      | VargElement<"music">
-      | VargElement<"audio">;
-    clipIndex: number; // first leaf clip of the container
-  }
-  const deferredAudioElements: DeferredAudio[] = [];
-
-  // ---------------------------------------------------------------------------
-  // flattenClip: recursively flatten nested clips into leaf clips.
-  //
-  // A "container clip" has child <Clip> elements. Its non-clip children
-  // (Speech, Music, Captions) are hoisted/deferred to span the container's
-  // time region, which starts at the first leaf clip's timeline position.
-  //
-  // A "leaf clip" has no child <Clip> elements — it contains visual/audio
-  // layers and is rendered directly by editly.
-  // ---------------------------------------------------------------------------
-  let clipIndexCounter = 0;
-
-  function flattenClip(clipElement: VargElement<"clip">): void {
-    const childClips: VargElement<"clip">[] = [];
-    const nonClipChildren: VargElement[] = [];
-
-    for (const child of clipElement.children) {
-      if (!child || typeof child !== "object" || !("type" in child)) continue;
-      const el = child as VargElement;
-      if (el.type === "clip") {
-        childClips.push(el as VargElement<"clip">);
-      } else {
-        nonClipChildren.push(el);
-      }
-    }
-
-    if (childClips.length === 0) {
-      // Leaf clip — hoist captions, keep everything else
-      const currentClipIndex = clipIndexCounter++;
-      const kept: typeof clipElement.children = [];
-      for (const el of nonClipChildren) {
-        if (el.type === "captions") {
-          hoistedCaptions.push({
-            element: el as VargElement<"captions">,
-            clipIndex: currentClipIndex,
-          });
-        } else {
-          kept.push(el);
-        }
-      }
-      clipElements.push({
-        ...clipElement,
-        children: kept,
-      } as VargElement<"clip">);
-      return;
-    }
-
-    // Container clip — has child clips.
-    // The first leaf clip's index is used as the anchor for audio/captions
-    // offsets (they all start at the container's position in the timeline).
-    const firstLeafClipIndex = clipIndexCounter; // before recursion increments it
-
-    // Collect overlays from container level — these get injected into each
-    // child clip so the overlay appears across all inner clips.
-    const containerOverlays: VargElement[] = [];
-    for (const el of nonClipChildren) {
-      if (el.type === "overlay") {
-        containerOverlays.push(el);
-      }
-    }
-
-    // Recurse into child clips, injecting container-level overlays
-    for (const childClip of childClips) {
-      if (containerOverlays.length > 0) {
-        childClip.children = [...childClip.children, ...containerOverlays];
-      }
-      flattenClip(childClip);
-    }
-
-    // Process remaining non-clip children at the container level
-    for (const el of nonClipChildren) {
-      if (el.type === "captions") {
-        hoistedCaptions.push({
-          element: el as VargElement<"captions">,
-          clipIndex: firstLeafClipIndex,
-        });
-      } else if (
-        el.type === "speech" ||
-        el.type === "music" ||
-        el.type === "audio"
-      ) {
-        deferredAudioElements.push({
-          element: el as
-            | VargElement<"speech">
-            | VargElement<"music">
-            | VargElement<"audio">,
-          clipIndex: firstLeafClipIndex,
-        });
-      }
-      // overlay: already handled above (distributed to child clips)
-      // Image/Video at container level: not supported yet (would need
-      // background layer spanning all child clips — a future feature)
-    }
-  }
-
-  // ---------------------------------------------------------------------------
-  // Process all children of <Render>
-  // ---------------------------------------------------------------------------
+  // Collect non-clip top-level children
+  const clipChildren: VargElement<"clip">[] = [];
   for (const child of element.children) {
     if (!child || typeof child !== "object" || !("type" in child)) continue;
     const childElement = child as VargElement;
 
     if (childElement.type === "clip") {
-      flattenClip(childElement as VargElement<"clip">);
+      clipChildren.push(childElement as VargElement<"clip">);
     } else if (childElement.type === "overlay") {
       overlayElements.push(childElement as VargElement<"overlay">);
     } else if (childElement.type === "captions") {
-      // Render-level captions — hoist with clipIndex 0 (start of timeline)
-      hoistedCaptions.push({
-        element: childElement as VargElement<"captions">,
-        clipIndex: 0,
-      });
+      // Render-level captions — hoist with clipIndex 0
+      // (handled in flatten result, but we need to add it manually)
     } else if (childElement.type === "speech") {
-      // Render-level speech — immediate audio track (no offset needed)
       const file =
         childElement instanceof ResolvedElement
           ? childElement.meta.file
@@ -425,7 +91,6 @@ export async function renderRoot(
         mixVolume: speechProps.volume ?? 1,
       });
     } else if (childElement.type === "audio") {
-      // Render-level derived audio node — immediate audio track
       const file = await renderAudio(childElement as VargElement<"audio">, ctx);
       const path = await ctx.backend.resolvePath(file);
       audioTracks.push({
@@ -437,18 +102,44 @@ export async function renderRoot(
     }
   }
 
-  // Hoisted captions are processed AFTER clip timeline offsets are computed
-  // (see below) so that each caption's audio can be delayed to the correct
-  // clip start time and ASS timestamps can be shifted accordingly.
+  // Flatten clip tree
+  const flatten = flattenClips(clipChildren);
 
+  // Add render-level captions (clipIndex 0)
+  for (const child of element.children) {
+    if (!child || typeof child !== "object" || !("type" in child)) continue;
+    const childElement = child as VargElement;
+    if (childElement.type === "captions") {
+      flatten.hoistedCaptions.push({
+        element: childElement as VargElement<"captions">,
+        clipIndex: 0,
+      });
+    }
+  }
+
+  return {
+    flatten,
+    overlayElements,
+    musicElements,
+    audioTracks,
+    hoistedCaptions: flatten.hoistedCaptions,
+  };
+}
+
+async function renderOverlays(
+  overlayElements: VargElement<"overlay">[],
+  ctx: ReturnType<typeof createRenderContext>["ctx"],
+): Promise<{ renderedOverlays: RenderedOverlay[]; audioTracks: AudioTrack[] }> {
   const renderedOverlays: RenderedOverlay[] = [];
+  const audioTracks: AudioTrack[] = [];
+
   for (const overlay of overlayElements) {
     const overlayProps = overlay.props as OverlayProps;
     for (const child of overlay.children) {
       if (!child || typeof child !== "object" || !("type" in child)) continue;
       const childElement = child as VargElement;
 
-      let file: File | undefined;
+      let file: import("../../ai-sdk/file").File | undefined;
       const isVideo = childElement.type === "video";
 
       if (childElement.type === "video") {
@@ -471,9 +162,33 @@ export async function renderRoot(
     }
   }
 
+  return { renderedOverlays, audioTracks };
+}
+
+export async function renderRoot(
+  element: VargElement<"render">,
+  options: RenderOptions,
+  prepared?: PreparedRender,
+): Promise<RenderResult> {
+  const props = element.props as RenderProps;
+  const { ctx, progress, mode, placeholderCount } =
+    prepared ?? createRenderContext(element, options);
+  const generatedFiles = ctx.generatedFiles;
+
+  // 1. Collect top-level children + flatten clip tree
+  const { flatten, overlayElements, musicElements, audioTracks } =
+    await collectTopLevel(element, ctx);
+
+  // 2. Render overlays
+  const { renderedOverlays, audioTracks: overlayAudio } = await renderOverlays(
+    overlayElements,
+    ctx,
+  );
+  audioTracks.push(...overlayAudio);
+
+  // 3. Render clips in parallel
   const concurrency =
     options.concurrency === undefined ? 3 : options.concurrency;
-
   if (
     concurrency !== Number.POSITIVE_INFINITY &&
     (!Number.isInteger(concurrency) || concurrency < 1)
@@ -482,7 +197,7 @@ export async function renderRoot(
   }
 
   const clipResults = await pMap(
-    clipElements,
+    flatten.clipElements,
     async (clipElement, i) => {
       try {
         return {
@@ -531,12 +246,13 @@ export async function renderRoot(
     return r.value;
   });
 
+  // 4. Assemble timeline with clip start offsets + overlay layers
   const clips: Clip[] = [];
   const clipStartOffsets: number[] = [];
   let currentTime = 0;
 
-  for (let i = 0; i < clipElements.length; i++) {
-    const clipElement = clipElements[i];
+  for (let i = 0; i < flatten.clipElements.length; i++) {
+    const clipElement = flatten.clipElements[i];
     const clip = renderedClips[i];
     if (!clipElement || !clip) {
       throw new Error(`Missing clip data at index ${i}`);
@@ -564,19 +280,18 @@ export async function renderRoot(
     clips.push(clip);
 
     currentTime += clipDuration;
-    if (i < clipElements.length - 1 && clip.transition) {
+    if (i < flatten.clipElements.length - 1 && clip.transition) {
       currentTime -= clip.transition.duration ?? 0;
     }
   }
 
   const totalDuration = currentTime;
 
-  // ---------------------------------------------------------------------------
-  // Process deferred audio from container clips.
-  // Now that clipStartOffsets are known, we can resolve the audio elements
-  // and set the correct start offset in the timeline.
-  // ---------------------------------------------------------------------------
-  for (const { element: audioElement, clipIndex } of deferredAudioElements) {
+  // 5. Process deferred audio from container clips
+  for (const {
+    element: audioElement,
+    clipIndex,
+  } of flatten.deferredAudioElements) {
     const offset = clipStartOffsets[clipIndex] ?? 0;
     if (audioElement.type === "speech") {
       const file =
@@ -622,16 +337,15 @@ export async function renderRoot(
     }
   }
 
-  // ---------------------------------------------------------------------------
-  // Process hoisted captions (from leaf clips, container clips, and Render
-  // level). Now that clipStartOffsets are known, each caption's audio can be
-  // delayed and ASS timestamps shifted to the correct timeline position.
-  // ---------------------------------------------------------------------------
+  // 6. Process hoisted captions
   const hoistedCaptionsResults: CaptionsResult[] = [];
   let mergedAssPath: string | undefined;
 
-  if (hoistedCaptions.length > 0) {
-    for (const { element: captionsElement, clipIndex } of hoistedCaptions) {
+  if (flatten.hoistedCaptions.length > 0) {
+    for (const {
+      element: captionsElement,
+      clipIndex,
+    } of flatten.hoistedCaptions) {
       const result = await renderCaptions(captionsElement, ctx);
       hoistedCaptionsResults.push(result);
 
@@ -646,7 +360,8 @@ export async function renderRoot(
 
     // Merge ASS files: shift timestamps by each clip's start offset
     if (hoistedCaptionsResults.length === 1) {
-      const offset = clipStartOffsets[hoistedCaptions[0]!.clipIndex] ?? 0;
+      const offset =
+        clipStartOffsets[flatten.hoistedCaptions[0]!.clipIndex] ?? 0;
       const assPath = hoistedCaptionsResults[0]!.assPath;
       mergedAssPath =
         offset > 0 ? shiftAssTimestamps(assPath, offset) : assPath;
@@ -656,7 +371,8 @@ export async function renderRoot(
     } else if (hoistedCaptionsResults.length > 1) {
       const segments = hoistedCaptionsResults.map((result, i) => ({
         assPath: result.assPath,
-        timeOffset: clipStartOffsets[hoistedCaptions[i]!.clipIndex] ?? 0,
+        timeOffset:
+          clipStartOffsets[flatten.hoistedCaptions[i]!.clipIndex] ?? 0,
         styleSuffix: `_${i}`,
       }));
       mergedAssPath = mergeAssFiles(segments, ctx.width, ctx.height);
@@ -664,7 +380,7 @@ export async function renderRoot(
     }
   }
 
-  // process music after clips so we know total duration for auto-trim
+  // 7. Process music after clips (need total duration for auto-trim)
   for (const musicElement of musicElements) {
     const musicProps = musicElement.props as MusicProps;
     const cutFrom = musicProps.cutFrom ?? 0;
@@ -672,7 +388,7 @@ export async function renderRoot(
       musicProps.cutTo ??
       (musicProps.duration !== undefined
         ? cutFrom + musicProps.duration
-        : totalDuration); // auto-trim to video length
+        : totalDuration);
 
     let path: string;
     if (musicProps.src) {
@@ -693,10 +409,8 @@ export async function renderRoot(
     });
   }
 
-  // Determine the ASS path to burn from merged hoisted captions.
-  const finalAssPath = mergedAssPath;
-  const hasCaptions = finalAssPath !== undefined;
-
+  // 8. Run editly
+  const hasCaptions = mergedAssPath !== undefined;
   const tempOutPath = hasCaptions
     ? `/tmp/varg-pre-captions-${Date.now()}.mp4`
     : (options.output ?? `output/varg-${Date.now()}.mp4`);
@@ -721,7 +435,8 @@ export async function renderRoot(
 
   let output = editlyResult.output;
 
-  if (hasCaptions && finalAssPath) {
+  // 9. Burn captions
+  if (hasCaptions && mergedAssPath) {
     const captionsTaskId = addTask(progress, "captions", "ffmpeg");
     startTask(progress, captionsTaskId);
 
@@ -738,7 +453,8 @@ export async function renderRoot(
     const allEmojiOverlays: EmojiOverlay[] = [];
     for (let i = 0; i < hoistedCaptionsResults.length; i++) {
       const result = hoistedCaptionsResults[i]!;
-      const offset = clipStartOffsets[hoistedCaptions[i]!.clipIndex] ?? 0;
+      const offset =
+        clipStartOffsets[flatten.hoistedCaptions[i]!.clipIndex] ?? 0;
       for (const overlay of result.emojiOverlays ?? []) {
         allEmojiOverlays.push(
           offset > 0
@@ -754,7 +470,7 @@ export async function renderRoot(
 
     output = await burnCaptions({
       video: output,
-      assPath: finalAssPath,
+      assPath: mergedAssPath,
       outputPath: finalOutPath,
       backend: options.backend,
       verbose: options.verbose,
@@ -769,6 +485,7 @@ export async function renderRoot(
     completeTask(progress, captionsTaskId);
   }
 
+  // 10. Read final output
   let finalBuffer: ArrayBuffer;
   if (output.type === "url") {
     const res = await fetch(output.url);
