+++ src/react/renderers/video.ts
@@ -1,87 +1,15 @@
 import { File } from "../../ai-sdk/file";
 import type { generateVideo } from "../../ai-sdk/generate-video";
 import { ResolvedElement } from "../resolved-element";
-import type {
-  ImageInput,
-  VargElement,
-  VideoPrompt,
-  VideoProps,
-} from "../types";
+import type { VargElement, VideoPrompt, VideoProps } from "../types";
 import type { RenderContext } from "./context";
-import { renderImage } from "./image";
-import { addTask, completeTask, startTask } from "./progress";
-import { renderSpeech } from "./speech";
-import { computeCacheKey, toFileUrl } from "./utils";
-
-async function resolveImageInput(
-  input: ImageInput,
-  ctx: RenderContext,
-): Promise<Uint8Array> {
-  if (input instanceof Uint8Array) {
-    return input;
-  }
-  if (typeof input === "string") {
-    const response = await fetch(toFileUrl(input));
-    return new Uint8Array(await response.arrayBuffer());
-  }
-  const file = await renderImage(input, ctx);
-  return file.arrayBuffer();
-}
-
-async function resolveAudioInput(
-  input:
-    | Uint8Array
-    | string
-    | VargElement<"speech">
-    | VargElement<"audio">
-    | undefined,
-  ctx: RenderContext,
-): Promise<Uint8Array | undefined> {
-  if (!input) return undefined;
-  if (input instanceof Uint8Array) return input;
-  if (typeof input === "string") {
-    const response = await fetch(toFileUrl(input));
-    return new Uint8Array(await response.arrayBuffer());
-  }
-  // Resolved speech/audio element — use pre-generated file directly
-  if (
-    input instanceof ResolvedElement &&
-    (input.type === "speech" || input.type === "audio")
-  ) {
-    return input.meta.file.arrayBuffer();
-  }
-  if (input.type === "speech") {
-    const file = await renderSpeech(input as VargElement<"speech">, ctx);
-    return file.arrayBuffer();
-  }
-  if (input.type === "audio") {
-    const { renderAudio } = await import("./audio");
-    const file = await renderAudio(input as VargElement<"audio">, ctx);
-    return file.arrayBuffer();
-  }
-  throw new Error(
-    `Unsupported audio input type: ${(input as VargElement).type}`,
-  );
-}
-
-async function resolveVideoInput(
-  input: Uint8Array | string | VargElement<"video"> | undefined,
-  ctx: RenderContext,
-): Promise<Uint8Array | undefined> {
-  if (!input) return undefined;
-  if (input instanceof Uint8Array) return input;
-  if (typeof input === "string") {
-    const response = await fetch(toFileUrl(input));
-    return new Uint8Array(await response.arrayBuffer());
-  }
-  if (input.type === "video") {
-    const file = await renderVideo(input, ctx);
-    return file.arrayBuffer();
-  }
-  throw new Error(
-    `Unsupported video input type: ${(input as VargElement).type}`,
-  );
-}
+import { withDedup } from "./dedup";
+import {
+  resolveAudioInput,
+  resolveImageInput,
+  resolveVideoInput,
+} from "./inputs";
+import { computeCacheKey } from "./utils";
 
 async function resolvePrompt(
   prompt: VideoPrompt,
@@ -164,22 +92,11 @@ export async function renderVideo(
     );
   }
 
+  const modelId = typeof model === "string" ? model : model.modelId;
   const cacheKey = computeCacheKey(element);
-  const cacheKeyStr = JSON.stringify(cacheKey);
 
-  const pendingRender = ctx.pendingFiles.get(cacheKeyStr);
-  if (pendingRender) {
-    return pendingRender;
-  }
-
-  const renderPromise = (async () => {
+  return withDedup(element, ctx, "video", modelId, async () => {
     const resolvedPrompt = await resolvePrompt(prompt, ctx);
-
-    const modelId = typeof model === "string" ? model : model.modelId;
-    const taskId = ctx.progress
-      ? addTask(ctx.progress, "video", modelId)
-      : null;
-    if (taskId && ctx.progress) startTask(ctx.progress, taskId);
 
     const { video } = await ctx.generateVideo({
       model,
@@ -189,8 +106,6 @@ export async function renderVideo(
       providerOptions: props.providerOptions,
       cacheKey,
     } as Parameters<typeof generateVideo>[0]);
-
-    if (taskId && ctx.progress) completeTask(ctx.progress, taskId);
 
     const mediaType = video.mimeType ?? "video/mp4";
     const promptText =
@@ -211,12 +126,6 @@ export async function renderVideo(
       await file.upload(ctx.storage);
     }
 
-    ctx.generatedFiles.push(file);
-
     return file;
-  })();
-
-  ctx.pendingFiles.set(cacheKeyStr, renderPromise);
-
-  return renderPromise;
+  });
 }
